+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Module doc links now end in `/index.html` instead of `/`
+- Only warn about missing readme when not `--strict(-crate-docs)`
 
 ## [0.3.0] - 2025-07-11